--- inputs/cipher_4/default.json
@@ -1,10 +1,0 @@
-{
-    "block": [
-      [50, 136, 49, 224],
-      [67, 90, 49, 55],
-      [246, 48, 152, 7],
-      [168, 141, 162, 52]
-    ],
-    "key": [43, 126, 21, 22, 40, 174, 210, 166, 171, 247, 21, 136, 9, 207, 79, 60]
-  }
-
